+++ modules/logging/ui/src/sap/commerce/toolset/logging/ui/CxBundledLogTemplatesView.kt
@@ -19,29 +19,17 @@
 package sap.commerce.toolset.logging.ui
 
 import com.intellij.ide.IdeBundle
-import com.intellij.ide.projectView.ProjectView
-import com.intellij.ide.util.PsiNavigationSupport
 import com.intellij.openapi.Disposable
-import com.intellij.openapi.application.edtWriteAction
-import com.intellij.openapi.application.invokeLater
-import com.intellij.openapi.application.readAction
 import com.intellij.openapi.observable.properties.AtomicBooleanProperty
-import com.intellij.openapi.observable.util.or
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogPanel
 import com.intellij.openapi.util.ClearableLazyValue
-import com.intellij.psi.PsiClass
-import com.intellij.psi.PsiPackage
-import com.intellij.psi.search.GlobalSearchScope
-import com.intellij.psi.util.startOffset
-import com.intellij.ui.EditorNotificationPanel
-import com.intellij.ui.InlineBanner
 import com.intellij.ui.components.JBScrollPane
-import com.intellij.ui.dsl.builder.*
+import com.intellij.ui.dsl.builder.Align
+import com.intellij.ui.dsl.builder.AlignX
+import com.intellij.ui.dsl.builder.TopGap
+import com.intellij.ui.dsl.builder.panel
 import com.intellij.util.ui.JBUI
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 import sap.commerce.toolset.logging.presentation.CxLoggerPresentation
 import java.awt.Dimension
 import javax.swing.JPanel
@@ -49,7 +37,6 @@ import javax.swing.JPanel
 class CxBundledLogTemplatesView(private val project: Project) : Disposable {
 
     private val showNothingSelected = AtomicBooleanProperty(true)
-    private val showNoLoggerTemplates = AtomicBooleanProperty(false)
     private val showDataPanel = AtomicBooleanProperty(false)
     private val initialized = AtomicBooleanProperty(true)
 
@@ -58,10 +45,9 @@ class CxBundledLogTemplatesView(private val project: Project) : Disposable {
         object : ClearableLazyValue<DialogPanel>() {
             override fun compute() = panel {
                 row {
-                    cellNoData(showNoLoggerTemplates, "No Logger Templates")
                     cellNoData(showNothingSelected, IdeBundle.message("empty.text.nothing.selected"))
                 }
-                    .visibleIf(showNoLoggerTemplates.or(showNothingSelected))
+                    .visibleIf(showNothingSelected)
                     .resizableRow()
                     .topGap(TopGap.MEDIUM)
 
@@ -100,7 +86,6 @@ class CxBundledLogTemplatesView(private val project: Project) : Disposable {
         get() = panel.value
 
     fun renderNothingSelected() = toggleView(showNothingSelected)
-    fun renderNoLoggerTemplates() = toggleView(showNoLoggerTemplates)
 
     fun renderLoggersTemplate(loggers: Map<String, CxLoggerPresentation>) {
         initialized.set(false)
@@ -109,7 +94,7 @@ class CxBundledLogTemplatesView(private val project: Project) : Disposable {
     }
 
     private fun renderLoggersInternal(loggers: Map<String, CxLoggerPresentation>) {
-        val view = if (loggers.isEmpty()) noLoggersView()
+        val view = if (loggers.isEmpty()) noLoggersView("No loggers configured for bundled Log Templates.")
         else createLoggersPanel(loggers.values)
 
         dataScrollPane.setViewportView(view)
@@ -120,75 +105,25 @@ class CxBundledLogTemplatesView(private val project: Project) : Disposable {
     private fun toggleView(vararg unhide: AtomicBooleanProperty) {
         listOf(
             showNothingSelected,
-            showNoLoggerTemplates,
             showDataPanel,
-            initialized
+            initialized,
         )
             .forEach { it.set(unhide.contains(it)) }
     }
 
-    private fun Row.cellNoData(property: AtomicBooleanProperty, text: String) = text(text)
-        .visibleIf(property)
-        .align(Align.CENTER)
-        .resizableColumn()
-
-    private fun noLoggersView() = panel {
-        row {
-            cell(
-                InlineBanner(
-                    "Unable to get list of loggers for the connection.",
-                    EditorNotificationPanel.Status.Warning
-                )
-                    .showCloseButton(false)
-            )
-                .align(Align.CENTER)
-                .resizableColumn()
-        }.resizableRow()
-    }
-
     fun createLoggersPanel(data: Collection<CxLoggerPresentation>) = panel {
-        data.forEach { r ->
+        data.forEach { cxLogger ->
             row {
-                icon(r.level.icon)
-                label(r.level.name)
+                icon(cxLogger.level.icon)
+                label(cxLogger.level.name)
                     .applyToComponent {
                         preferredSize = Dimension(JBUI.scale(68), preferredSize.height)
                     }
 
-                icon(r.icon)
-                if (r.resolved) {
-                    link(r.name) {
-                        r.psiElementPointer?.element?.let { psiElement ->
-                            when (psiElement) {
-                                is PsiPackage -> {
-                                    CoroutineScope(Dispatchers.Default).launch {
-                                        val directory = readAction {
-                                            psiElement.getDirectories(GlobalSearchScope.allScope(project))
-                                                .firstOrNull()
-                                        } ?: return@launch
-
-                                        edtWriteAction {
-                                            ProjectView.getInstance(project).selectPsiElement(directory, true)
-                                        }
-                                    }
-                                }
-
-                                is PsiClass -> invokeLater {
-                                    PsiNavigationSupport.getInstance()
-                                        .createNavigatable(project, psiElement.containingFile.virtualFile, psiElement.startOffset)
-                                        .navigate(true)
-                                }
-                            }
-                        }
-                    }.resizableColumn()
-                } else {
-                    label(r.name)
-                }
+                loggerName(project, cxLogger)
             }
         }
     }
 
-
     override fun dispose() = panel.drop()
-
 }